Aryan Laddha
There are 1010 problems in a contest. You know that the score of each problem is either 11 or 100100 points.

Chef came to know the total score of a participant and he is wondering how many problems were actually solved by that participant.

Given the total score PP of the participant, determine the number of problems solved by the participant. Print -1−1 in case the score is invalid.