Aryan Laddha
You have a grid with NN rows and MM columns. You have two types of tiles — one of dimensions 2 \times 22×2 and the other of dimensions 1 \times 11×1. You want to cover the grid using these two types of tiles in such a way that:

Each cell of the grid is covered by exactly one tile; and
The number of 1 \times 11×1 tiles used is minimized.
Find the minimum number of 1 \times 11×1 tiles you have to use to fill the grid.